// apps/web/lib/types.ts
export type PlanProps = "free" | "pro" | "enterprise";

export interface WorkspaceProps {
  id: string;
  name: string;
  studioKey: string;
  studioTokenEditor: string;
  slug: string;
  logo?: string;
  usage: number;
  usageLimit: number;
  plan: PlanProps;
  stripeId?: string;
  billingCycleStart?: number;
  createdAt?: Date;
  domains?: {
    slug: string;
  }[];
  users?: {
    role: "OWNER" | "MEMBER";
  }[];
}